Bugfixes, location, cleaning
Added POST of the user's location. In the process, I came across two bugs: one, if a string ever had a "." the querybuilder's protective measures would do funky things to it; two, changing multiple parameters where one went wrong disturbed the program flow to the point where it tried to send multiple messages (which of course isn't allowed). That's all fixed, as well as introducing a new class: UserInfoHandler, which automatically fills in the id of the user whose information is being queried.