Work Experience, Job Hunting
Reworked how work experience is represented in the data. Had to add two end points, change a few queries and more, but this should be a better fit for what was requested.

Also added the representation of jobhunting. It's not been worked in anywhere else, so that's something to work on.